Revert "Move tests from tests/pytestto tests"
This reverts commit e8e8a7ada613161f63844ffda32d5b89f9b2ed04.
