+++ ruff.toml
@@ -111,13 +111,13 @@ convention = "google"
 
 
 [lint.per-file-ignores]
-"test/*.py" = [
+"test/pytest/*.py" = [
     "INP001",  # `implicit-namespace-package`: test suite is not meant to be imported
 ]
-"test/test_*.py" = [
+"test/pytest/test_*.py" = [
     "S101",  # Allow uses of the `assert` keyword
 ]
-"test/Test*.py" = [
+"test/pytest/Test*.py" = [
     "S101",  # Allow uses of the `assert` keyword
 ]
 "docs/conf.py" = [
@@ -6184,29 +6184,31 @@ convention = "google"
 "mxcubecore/utils/units.py" = [
     "N802",
 ]
-"test/TestAbstractActuatorBase.py" = [
+"test/pytest/TestAbstractActuatorBase.py" = [
+    "EXE001",
     "N999",
     "PT012",
     "SLF001",
 ]
-"test/TestAbstractMotorBase.py" = [
+"test/pytest/TestAbstractMotorBase.py" = [
     "B017",
     "N999",
     "PT012",
     "SLF001",
 ]
-"test/TestAbstractNStateBase.py" = [
+"test/pytest/TestAbstractNStateBase.py" = [
     "B017",
     "C400",
     "N999",
     "PT012",
 ]
-"test/TestHardwareObjectBase.py" = [
+"test/pytest/TestHardwareObjectBase.py" = [
+    "EXE001",
     "N999",
     "PT022",
     "SLF001",
 ]
-"test/conftest.py" = [
+"test/pytest/conftest.py" = [
     "ERA001",
     "N814",
     "PT003",
@@ -6216,12 +6218,12 @@ convention = "google"
     "SLF001",
     "T201",
 ]
-"test/test_aperture.py" = [
+"test/pytest/test_aperture.py" = [
     "ERA001",
     "PT022",
     "SLF001",
 ]
-"test/test_base_hardware_objects.py" = [
+"test/pytest/test_base_hardware_objects.py" = [
     "ARG005",
     "B009",
     "B010",
@@ -6243,19 +6245,20 @@ convention = "google"
     "SIM118",
     "SLF001",
 ]
-"test/test_beam.py" = [
+"test/pytest/test_beam.py" = [
     "ERA001",
     "PERF401",
     "SLF001",
 ]
-"test/test_beam_definer.py" = [
+"test/pytest/test_beam_definer.py" = [
     "ERA001",
 ]
-"test/test_beamline_ho_id.py" = [
+"test/pytest/test_beamline_ho_id.py" = [
+    "EXE001",
     "PT022",
     "SIM300",
 ]
-"test/test_command_container.py" = [
+"test/pytest/test_command_container.py" = [
     "B009",
     "C408",
     "E501",
@@ -6276,65 +6279,70 @@ convention = "google"
     "SIM910",
     "SLF001",
 ]
-"test/test_detector.py" = [
+"test/pytest/test_detector.py" = [
+    "EXE001",
     "F841",
     "PIE808",
     "PT022",
 ]
-"test/test_detector_distance.py" = [
+"test/pytest/test_detector_distance.py" = [
+    "EXE001",
     "PT022",
 ]
-"test/test_energy.py" = [
+"test/pytest/test_energy.py" = [
     "ERA001",
     "PT022",
 ]
-"test/test_flux.py" = [
+"test/pytest/test_flux.py" = [
     "ERA001",
     "PT012",
     "PT022",
     "SLF001",
     "T201",
 ]
-"test/test_hwo_isara_maint.py" = [
+"test/pytest/test_hwo_isara_maint.py" = [
     "ARG002",
     "C408",
     "N802",
     "SLF001",
 ]
-"test/test_hwo_maxiv_ispyblims.py" = [
+"test/pytest/test_hwo_maxiv_ispyblims.py" = [
     "C408",
     "E501",
     "S105",
     "S106",
 ]
-"test/test_mach_info.py" = [
+"test/pytest/test_mach_info.py" = [
+    "EXE001",
     "PT022",
     "SLF001",
 ]
-"test/test_procedure.py" = [
+"test/pytest/test_procedure.py" = [
     "PIE804",
 ]
-"test/test_resolution.py" = [
+"test/pytest/test_resolution.py" = [
+    "PT011",
     "PT012",
     "PT022",
     "SLF001",
 ]
-"test/test_sample_view.py" = [
+"test/pytest/test_sample_view.py" = [
+    "EXE001",
     "PT022",
 ]
-"test/test_shutter.py" = [
+"test/pytest/test_shutter.py" = [
     "ERA001",
     "PT022",
     "T201",
 ]
-"test/test_transmission.py" = [
+"test/pytest/test_transmission.py" = [
     "ERA001",
     "PT022",
 ]
-"test/test_utils_units.py" = [
+"test/pytest/test_utils_units.py" = [
     "N802",
 ]
-"test/test_xrf.py" = [
+"test/pytest/test_xrf.py" = [
     "ERA001",
     "PT022",
     "S108",
