+++ test/pytest/TestAbstractActuatorBase.py
@@ -25,7 +25,7 @@ import abc
 import gevent
 import pytest
 
-from test import TestHardwareObjectBase
+from test.pytest import TestHardwareObjectBase
 
 test_object = TestHardwareObjectBase.test_object
