+++ test/pytest/data/epics_channels.yaml
