+++ test/pytest/data/exporter_channels.yaml
