+++ test/pytest/data/exporter_commands_channels.yaml
