+++ test/pytest/data/no_commands_channels.yaml
