+++ test/pytest/data/tango_channels.yaml
