+++ test/pytest/data/tango_commands_channels.yaml
