+++ test/pytest/data/tango_duo.yaml
