+++ test/pytest/test_beam.py
@@ -23,7 +23,7 @@
 import pytest
 
 from mxcubecore.HardwareObjects.abstract.AbstractBeam import BeamShape
-from test import TestHardwareObjectBase
+from test.pytest import TestHardwareObjectBase
 
 __copyright__ = """ Copyright © by MXCuBE Collaboration """
 __license__ = "LGPLv3+"
