+++ test/pytest/test_beam_definer.py
@@ -22,7 +22,7 @@
 
 import pytest
 
-from test.TestAbstractNStateBase import TestAbstractNStateBase
+from test.pytest.TestAbstractNStateBase import TestAbstractNStateBase
 
 __copyright__ = """ Copyright © by MXCuBE Collaboration """
 __license__ = "LGPLv3+"
