+++ test/pytest/test_detector.py
@@ -27,7 +27,7 @@ import math
 
 import pytest
 
-from test import TestHardwareObjectBase
+from test.pytest import TestHardwareObjectBase
 
 __copyright__ = """ Copyright © 2016 - 2020 by MXCuBE Collaboration """
 __license__ = "LGPLv3+"
