+++ test/pytest/test_energy.py
@@ -23,7 +23,7 @@ __license__ = "LGPLv3+"
 
 import pytest
 
-from test import TestAbstractActuatorBase
+from test.pytest import TestAbstractActuatorBase
 
 
 @pytest.fixture
