+++ test/pytest/test_resolution.py
@@ -21,7 +21,7 @@ hardware objects
 
 import pytest
 
-from test import TestAbstractMotorBase
+from test.pytest import TestAbstractMotorBase
 
 __copyright__ = """ Copyright © 2016 - 2022 by MXCuBE Collaboration """
 __license__ = "LGPLv3+"
