+++ test/pytest/test_setup_commands_channels.py
